refactor: improve store readability
