+++ apps/frontend/src/types/trip.ts
@@ -1,4 +1,18 @@
-import type { Location, Route } from './route';
+import type { Location } from './route';
+
+interface TripStoreActions {
+  setTitle: (title: string) => void;
+  setDescription: (description: string) => void;
+  setStartDate: (date: Date | null) => void;
+  setEndDate: (date: Date | null) => void;
+  setHasTripFinished: (finished: boolean) => void;
+  setStops: (stops: Array<{ id: number }>) => void;
+  setTransportMode: (mode: string) => void;
+  setLocations: (locations: Location[]) => void;
+  setTravelers: (
+    users: Array<{ id: number; email: string; role: string }>
+  ) => void;
+}
 
 export interface TripDetails {
   title: string;
@@ -22,4 +36,5 @@ export interface TripForm {
   details: TripDetails;
   route: TripRoute;
   travelers: TripTravelers;
+  actions: TripStoreActions;
 }
